+++ parsers.go
@@ -1,0 +1,68 @@
+package main
+
+import "encoding/json"
+
+func parseIssueComment(body []byte) (IssueComment, error) {
+	var message struct {
+		Issue struct {
+			Number      int `json:"Number"`
+			PullRequest struct {
+				URL string `json:"url"`
+			} `json:"pull_request"`
+		} `json:"issue"`
+		Repository struct {
+			Name  string `json:"name"`
+			Owner struct {
+				Login string `json:"login"`
+			} `json:"owner"`
+			SSHURL string `json:"ssh_url"`
+		} `json:"repository"`
+		Comment struct {
+			Body string `json:"body"`
+		} `json:"comment"`
+	}
+	err := json.Unmarshal(body, &message)
+	if err != nil {
+		return IssueComment{}, err
+	}
+	return IssueComment{
+		IssueNumber:   message.Issue.Number,
+		Comment:       message.Comment.Body,
+		IsPullRequest: message.Issue.PullRequest.URL != "",
+		Repository: Repository{
+			Owner: message.Repository.Owner.Login,
+			Name:  message.Repository.Name,
+			URL:   message.Repository.SSHURL,
+		},
+	}, nil
+}
+
+func parsePullRequestEvent(body []byte) (PullRequestEvent, error) {
+	var message struct {
+		Action      string `json:"action"`
+		Number      int    `json:"number"`
+		PullRequest struct {
+			URL string `json:"url"`
+		} `json:"pull_request"`
+		Repository struct {
+			Name  string `json:"name"`
+			Owner struct {
+				Login string `json:"login"`
+			} `json:"owner"`
+			SSHURL string `json:"ssh_url"`
+		} `json:"repository"`
+	}
+	err := json.Unmarshal(body, &message)
+	if err != nil {
+		return PullRequestEvent{}, err
+	}
+	return PullRequestEvent{
+		IssueNumber: message.Number,
+		Action:      message.Action,
+		Repository: Repository{
+			Owner: message.Repository.Owner.Login,
+			Name:  message.Repository.Name,
+			URL:   message.Repository.SSHURL,
+		},
+	}, nil
+}
